Add functionality to uninstall an NServiceBusHost
