+++ product/dropkick/Configuration/Dsl/NServiceBusHost/Action.cs
@@ -1,0 +1,15 @@
+﻿
+namespace dropkick.Configuration.Dsl.NServiceBusHost
+{
+    public enum Action
+    {
+        /// <summary>
+        /// Install the NServiceBus host
+        /// </summary>
+        Install,
+        /// <summary>
+        /// Uninstall the NServceBus host
+        /// </summary>
+        Uninstall
+    }
+}
