+++ product/dropkick/Tasks/NServiceBusHost/RemoteNServiceBusHostUninstallTask.cs
@@ -1,0 +1,57 @@
+﻿// Copyright 2007-2010 The Apache Software Foundation.
+// 
+// Licensed under the Apache License, Version 2.0 (the "License"); you may not use 
+// this file except in compliance with the License. You may obtain a copy of the 
+// License at 
+// 
+//     http://www.apache.org/licenses/LICENSE-2.0 
+// 
+// Unless required by applicable law or agreed to in writing, software distributed 
+// under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR 
+// CONDITIONS OF ANY KIND, either express or implied. See the License for the 
+// specific language governing permissions and limitations under the License.
+namespace dropkick.Tasks.NServiceBusHost
+{
+    using CommandLine;
+    using DeploymentModel;
+
+    public class RemoteNServiceBusHostUninstallTask :
+        BaseTask
+    {
+        readonly RemoteCommandLineTask _task;
+
+        public RemoteNServiceBusHostUninstallTask(string exeName, string location, string instanceName, PhysicalServer site, string serviceName)
+        {
+            string args = string.IsNullOrEmpty(instanceName)
+                            ? ""
+                            : " /instance:\"{0}\"".FormatWith(instanceName);
+
+            if (!string.IsNullOrEmpty(serviceName))
+                args += " /serviceName:\"{0}\"".FormatWith(serviceName);
+
+            _task = new RemoteCommandLineTask(exeName)
+            {
+                Args = "/install" + args,
+                ExecutableIsLocatedAt = location,
+                Machine = site.Name,
+                WorkingDirectory = location
+            };
+        }
+
+        public override string Name
+        {
+            get { return "[nservicebushost] remote Installing"; }
+        }
+
+        public override DeploymentResult VerifyCanRun()
+        {
+            return _task.VerifyCanRun();
+        }
+
+        public override DeploymentResult Execute()
+        {
+            Logging.Coarse("[nservicebushost] Installing a remote NServiceBus.Host service");
+            return _task.Execute();
+        }
+    }
+}
